requirements-dev.txt
pytest==8.3.4
pytest-cov==4.1.0
tabulate==0.9.0
loguru==0.7.3
ffmpeg-python==0.2.0